+++ src/translate/structure.rs
@@ -10,7 +10,7 @@ use unipn::PlaceId;
 /// Token capacity of an `RwLock` resource place. A read lock consumes one
 /// token (so up to `RWLOCK_CAPACITY` concurrent readers); a write lock is
 /// exclusive and must consume *all* tokens.
-pub const RWLOCK_CAPACITY: u64 = 10;
+pub const RWLOCK_CAPACITY: usize = 10;
 
 pub struct ResourceRegistry {
     locks: FxHashMap<AliasId, PlaceId>,
